requirements.txt
keyboard==0.13.5
pystray>=0.19.0
Pillow>=9.0.0